Refactored GRPC tests for queries and added test for boost parameter for all queries
