+++ src/test/java/org/apache/platypus/server/grpc/LuceneServerTest.java
@@ -256,365 +256,6 @@ public class LuceneServerTest {
         checkHits(secondHit);
     }
 
-    @Test
-    public void testSearchQueryText() throws IOException, InterruptedException {
-        GrpcServer.TestServer testAddDocs = new GrpcServer.TestServer(grpcServer, true, Mode.STANDALONE);
-        //2 docs addDocuments
-        testAddDocs.addDocuments();
-        //manual refresh
-        grpcServer.getBlockingStub().refresh(RefreshRequest.newBuilder().setIndexName(grpcServer.getTestIndex()).build());
-
-        SearchResponse searchResponse = grpcServer.getBlockingStub().search(SearchRequest.newBuilder()
-                .setIndexName(grpcServer.getTestIndex())
-                .setStartHit(0)
-                .setTopHits(10)
-                .addAllRetrieveFields(RETRIEVED_VALUES)
-                .setQueryText("SECOND")
-                .build());
-
-        assertEquals(1, searchResponse.getTotalHits());
-        assertEquals(1, searchResponse.getHitsList().size());
-        SearchResponse.Hit hit = searchResponse.getHits(0);
-        String docId = hit.getFieldsMap().get("doc_id").getFieldValue(0).getTextValue();
-        assertEquals("2", docId);
-        checkHits(hit);
-    }
-
-    @Test
-    public void testSearchBooleanQuery() throws IOException, InterruptedException {
-        GrpcServer.TestServer testAddDocs = new GrpcServer.TestServer(grpcServer, true, Mode.STANDALONE);
-        //2 docs addDocuments
-        testAddDocs.addDocuments();
-        //manual refresh
-        grpcServer.getBlockingStub().refresh(RefreshRequest.newBuilder().setIndexName(grpcServer.getTestIndex()).build());
-
-        SearchResponse searchResponse = grpcServer.getBlockingStub().search(SearchRequest.newBuilder()
-                .setIndexName(grpcServer.getTestIndex())
-                .setStartHit(0)
-                .setTopHits(10)
-                .addAllRetrieveFields(RETRIEVED_VALUES)
-                .setQuery(Query.newBuilder()
-                        .setQueryType(QueryType.BOOLEAN_QUERY)
-                        .setBooleanQuery(BooleanQuery.newBuilder()
-                                .addClauses(BooleanClause.newBuilder()
-                                        .setQuery(Query.newBuilder()
-                                                .setQueryType(QueryType.PHRASE_QUERY)
-                                                .setPhraseQuery(PhraseQuery.newBuilder()
-                                                        .setSlop(0)
-                                                        .setField("vendor_name")
-                                                        .addTerms("first")
-                                                        .addTerms("again")
-                                                        .build())
-                                                .build())
-                                        .setOccur(BooleanClause.Occur.MUST)
-                                        .build())
-                                .build())
-                        .build())
-                .build());
-
-        assertEquals(1, searchResponse.getTotalHits());
-        assertEquals(1, searchResponse.getHitsList().size());
-        SearchResponse.Hit hit = searchResponse.getHits(0);
-        String docId = hit.getFieldsMap().get("doc_id").getFieldValue(0).getTextValue();
-        assertEquals("1", docId);
-        checkHits(hit);
-    }
-
-    @Test
-    public void testSearchPhraseQuery() throws IOException, InterruptedException {
-        GrpcServer.TestServer testAddDocs = new GrpcServer.TestServer(grpcServer, true, Mode.STANDALONE);
-        //2 docs addDocuments
-        testAddDocs.addDocuments();
-        //manual refresh
-        grpcServer.getBlockingStub().refresh(RefreshRequest.newBuilder().setIndexName(grpcServer.getTestIndex()).build());
-
-        SearchResponse searchResponse = grpcServer.getBlockingStub().search(SearchRequest.newBuilder()
-                .setIndexName(grpcServer.getTestIndex())
-                .setStartHit(0)
-                .setTopHits(10)
-                .addAllRetrieveFields(RETRIEVED_VALUES)
-                .setQuery(Query.newBuilder()
-                        .setQueryType(QueryType.PHRASE_QUERY)
-                        .setPhraseQuery(PhraseQuery.newBuilder()
-                                .setSlop(0)
-                                .setField("vendor_name")
-                                .addTerms("second")
-                                .addTerms("again")
-                                .build())
-                        .build())
-                .build());
-
-        assertEquals(1, searchResponse.getTotalHits());
-        assertEquals(1, searchResponse.getHitsList().size());
-        SearchResponse.Hit hit = searchResponse.getHits(0);
-        String docId = hit.getFieldsMap().get("doc_id").getFieldValue(0).getTextValue();
-        assertEquals("2", docId);
-        checkHits(hit);
-    }
-
-    @Test
-    public void testSearchFunctionScoreQuery() throws IOException, InterruptedException {
-        GrpcServer.TestServer testAddDocs = new GrpcServer.TestServer(grpcServer, true, Mode.STANDALONE);
-        //2 docs addDocuments
-        testAddDocs.addDocuments();
-        //manual refresh
-        grpcServer.getBlockingStub().refresh(RefreshRequest.newBuilder().setIndexName(grpcServer.getTestIndex()).build());
-
-        SearchResponse searchResponse = grpcServer.getBlockingStub().search(SearchRequest.newBuilder()
-                .setIndexName(grpcServer.getTestIndex())
-                .setStartHit(0)
-                .setTopHits(10)
-                .addAllRetrieveFields(RETRIEVED_VALUES)
-                .setQuery(Query.newBuilder()
-                        .setQueryType(QueryType.FUNCTION_SCORE_QUERY)
-                        .setFunctionScoreQuery(FunctionScoreQuery.newBuilder()
-                                .setFunction("sqrt(4) * count")
-                                .setQuery(Query.newBuilder()
-                                        .setQueryType(QueryType.PHRASE_QUERY)
-                                        .setPhraseQuery(PhraseQuery.newBuilder()
-                                                .setSlop(0)
-                                                .setField("vendor_name")
-                                                .addTerms("second")
-                                                .addTerms("again")
-                                                .build()))
-                                .build())
-                        .build())
-                .build());
-
-        assertEquals(1, searchResponse.getTotalHits());
-        assertEquals(1, searchResponse.getHitsList().size());
-        SearchResponse.Hit hit = searchResponse.getHits(0);
-        String docId = hit.getFieldsMap().get("doc_id").getFieldValue(0).getTextValue();
-        assertEquals("2", docId);
-        assertEquals(14.0, hit.getScore(), 0.0);
-        checkHits(hit);
-    }
-
-    @Test
-    public void testSearchTermQuery() throws IOException, InterruptedException {
-        GrpcServer.TestServer testAddDocs = new GrpcServer.TestServer(grpcServer, true, Mode.STANDALONE);
-        //2 docs addDocuments
-        testAddDocs.addDocuments();
-        //manual refresh
-        grpcServer.getBlockingStub().refresh(RefreshRequest.newBuilder().setIndexName(grpcServer.getTestIndex()).build());
-
-        SearchResponse searchResponse = grpcServer.getBlockingStub().search(SearchRequest.newBuilder()
-                .setIndexName(grpcServer.getTestIndex())
-                .setStartHit(0)
-                .setTopHits(10)
-                .addAllRetrieveFields(RETRIEVED_VALUES)
-                .setQuery(Query.newBuilder()
-                        .setQueryType(QueryType.TERM_QUERY)
-                        .setTermQuery(TermQuery.newBuilder()
-                                .setField("vendor_name")
-                                .setTerm("second")))
-                .build());
-
-        assertEquals(1, searchResponse.getTotalHits());
-        assertEquals(1, searchResponse.getHitsList().size());
-        SearchResponse.Hit hit = searchResponse.getHits(0);
-        String docId = hit.getFieldsMap().get("doc_id").getFieldValue(0).getTextValue();
-        assertEquals("2", docId);
-        checkHits(hit);
-    }
-
-    @Test
-    public void testSearchTermInSetQuery() throws IOException, InterruptedException {
-        GrpcServer.TestServer testAddDocs = new GrpcServer.TestServer(grpcServer, true, Mode.STANDALONE);
-        //2 docs addDocuments
-        testAddDocs.addDocuments();
-        //manual refresh
-        grpcServer.getBlockingStub().refresh(RefreshRequest.newBuilder().setIndexName(grpcServer.getTestIndex()).build());
-
-        SearchResponse searchResponse = grpcServer.getBlockingStub().search(SearchRequest.newBuilder()
-                .setIndexName(grpcServer.getTestIndex())
-                .setStartHit(0)
-                .setTopHits(10)
-                .addAllRetrieveFields(RETRIEVED_VALUES)
-                .setQuery(Query.newBuilder()
-                        .setQueryType(QueryType.TERM_IN_SET_QUERY)
-                        .setTermInSetQuery(TermInSetQuery.newBuilder()
-                                .setField("vendor_name")
-                                .addTerms("second")
-                                .addTerms("first")))
-                .build());
-
-        assertEquals(2, searchResponse.getTotalHits());
-        assertEquals(2, searchResponse.getHitsList().size());
-        SearchResponse.Hit firstHit = searchResponse.getHits(0);
-        checkHits(firstHit);
-        SearchResponse.Hit secondHit = searchResponse.getHits(1);
-        checkHits(secondHit);
-    }
-
-    @Test
-    public void testSearchDisjunctionMaxQuery() throws IOException, InterruptedException {
-        GrpcServer.TestServer testAddDocs = new GrpcServer.TestServer(grpcServer, true, Mode.STANDALONE);
-        //2 docs addDocuments
-        testAddDocs.addDocuments();
-        //manual refresh
-        grpcServer.getBlockingStub().refresh(RefreshRequest.newBuilder().setIndexName(grpcServer.getTestIndex()).build());
-
-        SearchResponse searchResponse = grpcServer.getBlockingStub().search(SearchRequest.newBuilder()
-                .setIndexName(grpcServer.getTestIndex())
-                .setStartHit(0)
-                .setTopHits(10)
-                .addAllRetrieveFields(RETRIEVED_VALUES)
-                .setQuery(Query.newBuilder()
-                        .setQueryType(QueryType.DISJUNCTION_MAX)
-                        .setDisjunctionMaxQuery(DisjunctionMaxQuery.newBuilder()
-                                .addDisjuncts(Query.newBuilder()
-                                        .setQueryType(QueryType.TERM_QUERY)
-                                        .setTermQuery(TermQuery.newBuilder()
-                                                .setField("vendor_name")
-                                                .setTerm("second")))
-                                .addDisjuncts(Query.newBuilder()
-                                        .setQueryType(QueryType.FUNCTION_SCORE_QUERY)
-                                        .setFunctionScoreQuery(FunctionScoreQuery.newBuilder()
-                                                .setFunction("sqrt(4) * count")
-                                                .setQuery(Query.newBuilder()
-                                                        .setQueryType(QueryType.TERM_QUERY)
-                                                        .setTermQuery(TermQuery.newBuilder()
-                                                                .setField("vendor_name")
-                                                                .setTerm("second")))))
-                                .setTieBreakerMultiplier(0)))
-                .build());
-
-        assertEquals(1, searchResponse.getTotalHits());
-        assertEquals(1, searchResponse.getHitsList().size());
-        SearchResponse.Hit hit = searchResponse.getHits(0);
-        String docId = hit.getFieldsMap().get("doc_id").getFieldValue(0).getTextValue();
-        assertEquals("2", docId);
-        assertEquals(14.0, hit.getScore(), 0.0);
-        checkHits(hit);
-    }
-
-    @Test
-    public void testSearchMatchQuery() throws IOException, InterruptedException {
-        GrpcServer.TestServer testAddDocs = new GrpcServer.TestServer(grpcServer, true, Mode.STANDALONE);
-        //2 docs addDocuments
-        testAddDocs.addDocuments();
-        //manual refresh
-        grpcServer.getBlockingStub().refresh(RefreshRequest.newBuilder().setIndexName(grpcServer.getTestIndex()).build());
-
-        SearchResponse searchResponse = grpcServer.getBlockingStub()
-                .search(SearchRequest.newBuilder()
-                        .setIndexName(grpcServer.getTestIndex())
-                        .setStartHit(0)
-                        .setTopHits(10)
-                        .addAllRetrieveFields(RETRIEVED_VALUES)
-                        .setQuery(Query.newBuilder()
-                                .setQueryType(QueryType.MATCH)
-                                .setMatchQuery(MatchQuery.newBuilder()
-                                        .setField("vendor_name")
-                                        .setQuery("SECOND again")
-                                        .setOperator(MatchOperator.MUST)))
-                        .build());
-
-        assertEquals(1, searchResponse.getTotalHits());
-        assertEquals(1, searchResponse.getHitsList().size());
-        SearchResponse.Hit hit = searchResponse.getHits(0);
-        String docId = hit.getFieldsMap().get("doc_id").getFieldValue(0).getTextValue();
-        assertEquals("2", docId);
-        checkHits(hit);
-    }
-
-    @Test
-    public void testSearchMatchQueryFuzzyCustomAnalyzer() throws IOException, InterruptedException {
-        GrpcServer.TestServer testAddDocs = new GrpcServer.TestServer(grpcServer, true, Mode.STANDALONE);
-        //2 docs addDocuments
-        testAddDocs.addDocuments();
-        //manual refresh
-        grpcServer.getBlockingStub().refresh(RefreshRequest.newBuilder().setIndexName(grpcServer.getTestIndex()).build());
-
-        SearchResponse searchResponse = grpcServer.getBlockingStub()
-                .search(SearchRequest.newBuilder()
-                        .setIndexName(grpcServer.getTestIndex())
-                        .setStartHit(0)
-                        .setTopHits(10)
-                        .addAllRetrieveFields(RETRIEVED_VALUES)
-                        .setQuery(Query.newBuilder()
-                                .setQueryType(QueryType.MATCH)
-                                .setMatchQuery(MatchQuery.newBuilder()
-                                        .setField("vendor_name")
-                                        .setQuery("<br> SEND agn </br>")
-                                        .setFuzzyParams(FuzzyParams.newBuilder()
-                                                .setMaxEdits(2)
-                                                .setPrefixLength(2)
-                                                .setMaxExpansions(1))
-                                        .setAnalyzer(getTestAnalyzer())
-                                        .setOperator(MatchOperator.MUST)))
-                        .build());
-
-        assertEquals(1, searchResponse.getTotalHits());
-        assertEquals(1, searchResponse.getHitsList().size());
-        SearchResponse.Hit hit = searchResponse.getHits(0);
-        String docId = hit.getFieldsMap().get("doc_id").getFieldValue(0).getTextValue();
-        assertEquals("2", docId);
-        checkHits(hit);
-    }
-
-    @Test
-    public void testSearchMatchPhraseQuery() throws IOException, InterruptedException {
-        GrpcServer.TestServer testAddDocs = new GrpcServer.TestServer(grpcServer, true, Mode.STANDALONE);
-        //2 docs addDocuments
-        testAddDocs.addDocuments();
-        //manual refresh
-        grpcServer.getBlockingStub().refresh(RefreshRequest.newBuilder().setIndexName(grpcServer.getTestIndex()).build());
-
-        SearchResponse searchResponse = grpcServer.getBlockingStub()
-                .search(SearchRequest.newBuilder()
-                        .setIndexName(grpcServer.getTestIndex())
-                        .setStartHit(0)
-                        .setTopHits(10)
-                        .addAllRetrieveFields(RETRIEVED_VALUES)
-                        .setQuery(Query.newBuilder()
-                                .setQueryType(QueryType.MATCH_PHRASE)
-                                .setMatchPhraseQuery(MatchPhraseQuery.newBuilder()
-                                        .setField("vendor_name")
-                                        .setQuery("SECOND second")
-                                        .setSlop(1)))
-                        .build());
-
-        assertEquals(1, searchResponse.getTotalHits());
-        assertEquals(1, searchResponse.getHitsList().size());
-        SearchResponse.Hit hit = searchResponse.getHits(0);
-        String docId = hit.getFieldsMap().get("doc_id").getFieldValue(0).getTextValue();
-        assertEquals("2", docId);
-        checkHits(hit);
-    }
-
-    @Test
-    public void testSearchMatchPhraseQueryCustomAnalyzer() throws IOException, InterruptedException {
-        GrpcServer.TestServer testAddDocs = new GrpcServer.TestServer(grpcServer, true, Mode.STANDALONE);
-        //2 docs addDocuments
-        testAddDocs.addDocuments();
-        //manual refresh
-        grpcServer.getBlockingStub().refresh(RefreshRequest.newBuilder().setIndexName(grpcServer.getTestIndex()).build());
-
-        SearchResponse searchResponse = grpcServer.getBlockingStub()
-                .search(SearchRequest.newBuilder()
-                        .setIndexName(grpcServer.getTestIndex())
-                        .setStartHit(0)
-                        .setTopHits(10)
-                        .addAllRetrieveFields(RETRIEVED_VALUES)
-                        .setQuery(Query.newBuilder()
-                                .setQueryType(QueryType.MATCH_PHRASE)
-                                .setMatchPhraseQuery(MatchPhraseQuery.newBuilder()
-                                        .setField("vendor_name")
-                                        .setQuery("<br> SECOND again </br>")
-                                        .setAnalyzer(getTestAnalyzer())
-                                        .setSlop(1)))
-                        .build());
-
-        assertEquals(1, searchResponse.getTotalHits());
-        assertEquals(1, searchResponse.getHitsList().size());
-        SearchResponse.Hit hit = searchResponse.getHits(0);
-        String docId = hit.getFieldsMap().get("doc_id").getFieldValue(0).getTextValue();
-        assertEquals("2", docId);
-        checkHits(hit);
-    }
-
     public static void checkHits(SearchResponse.Hit hit) {
         Map<String, CompositeFieldValue> fields = hit.getFieldsMap();
         List<String> expectedRetrieveFields = RETRIEVED_VALUES;
@@ -670,16 +311,6 @@ public class LuceneServerTest {
         return fieldValues.stream()
                 .map(SearchResponse.Hit.FieldValue::getTextValue)
                 .collect(Collectors.toList());
-    }
-
-    private Analyzer getTestAnalyzer() {
-        return Analyzer.newBuilder()
-                .setCustom(CustomAnalyzer.newBuilder()
-                        .setDefaultMatchVersion("LATEST")
-                        .addCharFilters(NameAndParams.newBuilder().setName("htmlstrip"))
-                        .setTokenizer(NameAndParams.newBuilder().setName("standard"))
-                        .addTokenFilters(NameAndParams.newBuilder().setName("lowercase")))
-                .build();
     }
 
 }
